📃 chore: 型定義の場所を適切にした
一度しか使用されていない型定義は使用する各ファイルに移動、使用していない型は削除。
