--- src/types/editor.ts
@@ -1,6 +1,0 @@
-import { Heading } from './heading';
-
-export type EditorProps = {
-  heading: Heading | undefined;
-  handleSave: (content: string, title: string) => void;
-};
