+++ src/types/index.ts
@@ -2,7 +2,3 @@ export * from './book';
 export * from './heading';
 export * from './memo';
 export * from './log';
-export * from './user';
-export * from './feature';
-export * from './item';
-export * from './editor';
