+++ src/types/log.ts
@@ -2,7 +2,3 @@ export type Log = {
   date: string;
   count: number;
 };
-
-export type ReadingLogs = {
-  readingLogs: Log[];
-};
