--- src/types/user.ts
@@ -1,11 +1,0 @@
-export type UserInfo = {
-  name: string;
-  id: string;
-  image: string;
-};
-
-export type UserParams = {
-  id: string;
-  token?: string;
-  close?: () => void;
-};
